requirements.txt
annotated-types==0.6.0
astroid==3.0.3
beautifulsoup4==4.12.3
bs4==0.0.2
dill==0.3.8
dnspython==2.5.0
exceptiongroup==1.2.0
iniconfig==2.0.0
isort==5.13.2
loguru==0.7.2
mccabe==0.7.0
mypy==1.8.0
mypy-extensions==1.0.0
packaging==23.2
platformdirs==4.2.0
pluggy==1.4.0
pycountry==23.12.11
pydantic==2.6.1
pydantic_core==2.16.2
pylint==3.0.3
pymongo==4.6.1
pytest==8.0.0
python-dateutil==2.8.2
six==1.16.0
soupsieve==2.5
tomli==2.0.1
tomlkit==0.12.3
tqdm==4.66.2
types-beautifulsoup4==4.12.0.20240106
types-html5lib==1.1.11.20240217
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
python-dotenv~=1.0.1